feat(logging): extract color logger to separate module
- Moved ColorLogger and ColorFormatter classes from node_sync_test.py
  to a new module color_logger.py
- Updated imports in node_sync_test.py to use the new color_logger module
